refactor `ExperimentSetup` into separate objects, allow change via cfg
Some of the new types can be changed via the config.toml file.

This is a pretty huge refactor. Still some stuff to do, as the
reflectivity code is still not done. The rest at least compiles again
now.

Possibly we might want to split the `ExperimentSetup` fully into
separate objects in the body of `traceAxion` to get rid of all the
multiple indirections. Probably they don't matter, but if we see a
performance regression that's something to check.

Furtehr: there is still plenty of stuff that needs to be refactored out.