refactor out a `lineIntersectsOpaqueTelescopeStructure` from code
This is not fully done and maybe somewhat of a hack. The code was
almost completely duplicated, but I don't fully understand why / what
the intention is. The `if` check didn't exist in the latter case,
hence the addition of the `onlyInnerPart`.

Also this needs to become dependent on the telescope and not the
experiment setup.