remove fake 0 layer for LLNL telescope
There was an additional "layer" corresponding to the opaque structure
holding the shells in place used for the LLNL telescope. As such there
were actually N+1 elements in the layer relevant sequences. But this
was not done as such in the other telescope.

Remove it, and instead handle `hitLayer` correctly (see also previous
and next commit).