change order of nested tables, make ref tables
Without the tables being references, accessing the element causes a
copy of the data. That can be expensive. For successive accesses
without assigning to a variables, that is optimized out, but if we
want to lift out the table before the Z loop, we need a ref so that
the operation is cheap.