[raytracer] add effective area scan & X-ray source using Sun distr.
The X-ray source can now also be used together with the Sun as an
emission source. The idea is to be able to use the geometric
properties as the Sun as a source, but use a fixed energy instead of
also sampling the energy.

This is the basis for computing the effective area for a realistic
emission model.

The effective area scan walks all given input energies and computes
the mean of the photon weights after tracing. For that all additional
factors (conversion probability, detector window, absorption) are
ignored of course and only the telescope reflectivity (and obv. loss
of X-rays completely) are included.